Block cache updates in the service worker if the server isn't up to date
This allows us to deploy new server & app versions independently.
We just have to keep the server backward compatible, and then
always release it first, so it can update itself, and then the
app will automatically update itself.

In future we can also tighten these contraints. If we want a breaking
server release, we need to break the version and find a way to
orchestrate updates of that for users (maybe just say 'update required',
then update both when they agree?). Should be able to put it off for a
while anyway.