Support fragment queries in OpenAPI specs
Some OpenAPI specs include paths like /123/456#a&b. These match
requests to /123/456 that include an 'a' and 'b' parameter.
They're specified as fragments because those are legal (and
generally ignored) in OpenAPI paths, where ? is not.

This is mainly of use in AWS, where many of the APIs are not
really rest, and use exactly the same path for many different
operations, but parameters like ?Action=CopyObject to define
what operation exactly is going to be executed.