Update to workbox v4, with lots of related SW fixes & improvements
This is a semi-optimistic fix for persistent issues with breaking caches.
The cache failure logic has been simplified and is less hacky, and the
change to workbox 4 and move to a manual replacement for the precache
lookup logic will hopefully avoid these issues in future, or at least
make the cause a little clearer.

This change will invalidate & replace all existing SW caches. That's
safe only because all current deployments support all existing servers.
This will briefly double the amount of space caches take up, but it'll
drop the previous caches entirely when the new SW is activated fully.