Fix SW, by always giving the precache controller the file list
Previously it was only provided on initial install. This worked fine for
the install and subsequent activation, but if the service worker later
stops, and then starts again afresh, it loses this state completely,
and never recovers. It doesn't read cache state from the cache itself,
only by explicitly providing it, so this means future requests are
never matched, and always fail