Reorganize a clearer separation of model/service-communication logic
Model here describes fundamental parts of the practical model of how
HTTP Toolkit sees the world. These are not pure implementation details,
they're descriptions of the data and how it's used and interpreted.
Meanwhile the services describe how the data processing works, moving
some logic execution into workers, or communicating with the other
desktop/server parts of the app to coordinate with them.