Avoid paddle's mobile export bug
This same issue happens in httptoolkit.tech. If Paddle
detects that its running on mobile (sometimes seems to
happen) then it loads differently. It uses some code
that includes a UMD header (unlike the rest of the
logic), which then defines itself as a module instead
of a global. The rest of the code assumes everything
is global, and therefore breaks.