Silence ResizeObserver loop warning errors
These are triggered if at any point ResizeObserver can't deliver
a resize event. Typically if a resize event triggers another DOM
change and another resize event in turn, but I think also if
there simply isn't time to handle it in the current frame or
similar. It has no visual effect though, as we don't depend on
immediate delivery of these events, so it's fine to ignore.

https://stackoverflow.com/questions/49384120/resizeobserver-loop-limit-exceeded
https://github.com/WICG/ResizeObserver/issues/38