Avoid Chrome's navigation throttling by wrapping pushState
Chrome throttles navigation events if they happen too quickly,
blocking all of them for a cooloff period. This can happen if
scrolling rapidly (e.g. hold the down arrow) on the events
list, and results in blocking any UI navigation. This PR fixes
that, by wrapping pushState to throttle updates to the history
itself, whilst still providing instant feedback & updates on
top with an in-memory store.