Persist exchange details card collapse state
Previously, it was only persisted whilst the details were
shown, so reset if you went to another page, deselected
all exchanges entirely, or looked at a TLS event.

It's now persisted in the UI store, so persists during
the entire app session, reseting each time the app is
opened/refreshed.