Add an optional auth token to further control API access
The token will be generated fresh by the desktop on every start, and
passed to both the server & the UI, so they can verify each other.

Only applies to the HTK server for now, not Mockttp, but that's
the most risky attack surface by a mile. No immediate need for
this - the CORS & CSRF protections are rock-solid AFAIK, but
as the API gets more powerful it would be good to have defence
in depth.