WIP: Update monaco, to drop workaround for showing readonly markers
This isn't usable yet, as it can crash the tab due to
https://github.com/microsoft/monaco-editor/issues/2014. Once that's
resolved, we might be able to continue this.

To reproduce this issue, run HTTP Toolkit, open Chrome, force refresh
the page so that the raw font file from fonts.gstatic.com is loaded,
view that response, then click away from the response -> crash.
In some cases you might need to shrink/expand the editor a bit to
make the word wrapping hit the trigger for this.