Fix mapRules when used with array-type results
Previously normal result values each become an element in
the resulting array, but array values were flattened as a
series of elements (rather than adding a single array
element into the final results array). This is because
concat has a special case for arrays - to avoid this, we
now explicitly wrap all values as arrays.