Ensure the view list can be focused when the selected row is hidden
Without this you can select a row, scroll away from it, and then you're
unable to tab to the list at all, which is a bit messy. Now you can tab
to the container directly in that case, and up/down from there still act
as you'd expect.

Briefly investigated scrolling automatically to the correct row in that
case, but difficult to do without disrupting focus behaviour in other
cases, such as mouse selection of rows. Simpler is reliabler.