Ensure list filtering doesn't get broken by the list focus updates
This was just reintroduced by a very recent commit - keeping it as a separate
fixed rather than fixing up to make it clearer why this logic exists. Without
this, typing in the filter box updates the list, which triggers didUpdate,
which moves the focus out of the filter box. This fixes that.