Improve text encoding handling when editing binary data
Previously, data was always encoded as UTF8. If you tried to edit
any mock rule with a body containing non-UTF8 data, that was a
lossy operation, where some of the body would be changed incorrectly.
Now, we only decode as UTF8 when the data really is UTF8, or when
we're just reading the data. When editing non-UTF8 data, we use
binary encoding (latin1) instead, since it's lossy.

This doesn't perfectly solve the problem, as monaco isn't really
built for editing binary data, but it solves it for all edits
to rule/breakpointed message metadata, and mostly solves editing
binary bodies too, or at least reduces the damage.