Fix force-update detection with new servers
Very recent server releases now reject non-POST requests,
_before_ checking the auth token. That means that this
auth connectivity check always returns 'accessible', and
we never force an update.

This comes up if a user uses a new desktop release on top
of an old UI. In that case, because we don't detect the
auth issue, we don't force update the cache. Then,
because getVersion never completes (due to graphql.js
bug in previous commit) the existing SW never installs,
and we never update.