Improve force-update protections by explicitly checking for a SW
If there's an active service worker, then something may (is almost certainly)
be cached, and that's exactly what we're fixing. As soon as the force-update
SW has triggered, the page will refresh with no service worker at all, loading
everything from scratch, and in that case there'll be no cache, and no service
worker.

This doesn't make a huge difference, but does simplify the code, and ensures
we don't bother with force update loops when there's no SW at all, in some
kind of situation where we're simply totally broken (no token available).