Don't cache the google-fonts bundle
This file doesn't seem to exist. That messes up local SW dev,
and causes the prod SW to cache the index.html page (presumably
because Netlify serves that for direct requests). I strongly
suspect this is a temp file created by the fonts plugin. The
real fonts are stored inside fonts.css, which does seem to
be cached correctly no problem.