Ensure suggestions are always updated correctly
Previously, clicking a suggestion (but not selecting it via
keyboard) wouldn't properly update the suggestion, due to
async updates as the input moved around. All of that happens
because we're dependent on react-autosuggest to tell us when
to update the list. We don't need to do that, so instead we
now update the list every time the text changes, and thus
guarantee that it's always up to date.