Fix bug when clicking filter suggestions
Clicking anything but the first suggestion did not select the
suggestion, because it triggered a selected filter tags update,
which wasn't pure - it set selectedFilters differently and
fired a new update every time, which broke the input click
process. Comparing structurally is cheap (filters are tiny
and the comparison only happens whilst filters are selected)
and neatly solves the problem by silencing unnecessary updates.