Explicitly suggest empty string values where valid
Previously we just suggested the template, but allowed an empty template.
Now we explicitly suggest both the template & empty string, with separate
descriptions for each. This necesitated extending filterDescripton so
it's aware of templates and making allowEmpty a function so we can limit
the operators that are allowed empty values (and avoid having to write
or show descriptions that don't make any sense).