Improve handling of filter suggestion prefix matches (again)
We'd improved this somewhat by constraining suggestions used in full
matches, but the problem is actually more general: when options saw a
full match, it returned _all_ full matches including prefixes. Instead
options should always pick the longest match: if you've entered a longer
exact match for an option, that's clearly the one you wanted. Similarly
(as before) if you've entered an exact match for a shorter option, we
should assume that that's correct.