Import & export TLS errors to HAR
This exports to a new custom field: _tlsError. The format closely
matches the existing HAR entries format, but without the HTTP-specific
details.

While this requires a server mockttp update & release, it's still enabled
on old releases anyway - all data will be correct, except starttime,
which will be set to the moment export for all such errors, which seems
reasonable.