Import HAR headers as lowercase
Semantically, it makes no difference but a) this better matches
what we do in Mockttp (which effectively lowercases headers
before reporting them) and b) this avoids some bugs - in a few
places, notably around content-types we assume headers are
lowercase. That's a very convenient assumption to keep, which
is safe with Mockttp (and is spec-safe in HTTP/2) so this
is cleaner than carefully fixing all of that.