Fix HAR export of client errors
Some client errors were not exported correctly, with their
wait timing set to null. In fact, this was a NaN, because
we assumed that having a headers-sent timestamp implied we
would have a body-received timestamp. That's not necessarily
true, if there's a parse error before the body turns up
(and probably in some other cases too) so we now catch that.