Improve OpenAPI dereferencing performance
This matters with the updated OpenAPI specs, where Stripe now includes
(near-?)endless loops of $refs. There's no need to dereference inside a
ref though, since we'll always traverse it in its initial position
anyway, so we can skip those cases.

Stripe is still a bit slower than I'd like here (a couple of seconds to
process) but it's on a background thread, and only affects Stripe usage,
so probably not a huge deal for now.