Don't use query params in OpenAPI parameter parsing
Unclear why, but Mockttp includes search params in its request.path
value. We shouldn't be using that anyway, as we have a parsedUrl
property which carefully separates the URL parts and that's used in
other param parsing too.

This manifests with OpenAPI requests for operations like /abc/{param},
which automatically include the full query string in the param value.

We now use the correct property, and req.path is removed from the types
so we can't use it accidentally in future.