Add target=_blank to add markdown links
The Electron shell should capture all these navigations regardless,
but it's good to be extra safe here, and it does catch one extra case:
links to 'about:blank', which incorrectly fail to fire will-navigate
events in Electron, so don't get intercepted. Better to open a new
window, so Electron can block & report it.