Fix HAR & code export of urlencoded request bodies
This was broken, because we always encoded bodies as text in HARs
but simultaneously passed params: []. Instead, we need to either
set text or params, and leave the other undefined. Because of
this, httpsnippet generated code for requests that used the params
value instead of the text value, and so included no body at all.

We now remove params entirely when only text content is available
and use only params with properly parsed param data when we do
have URL encoded form data, to ensure exported code uses friendly
arguments, instead of raw text, where possible.