Fix wrapper handling of multi-part suggestions
Previously, the wrapper assumed the suggestion ways always placed
at the start of the value it was suggesting for. That's not
necessarily true, especially with multi-part content, where it's
likely a later part will provide suggestions for its own part of
the value, rather than the entire thing.

With the fledgling Or(..., ...) implementation, this resulted in
suggestions like POST for Or(method= creating Or(POST, which is
clearly not quite right.

We now backtrack when required in these cases, to realign
suggestions in the cases where that's actually necessary.