Split out zstd from bundle and trim its content a little
The expectation is that very few users will actually need zstd, so
there's no point in loading it until it's actually needed. We do still
download it as part of prefetching in the service worker, so that this
works offline etc, but off the critical path of app startup, and it's
not actually executed until later.

Not necessary for Brotli, even though it's also not used that much,
because it's an async wasm module now so it's split out automatically
anyway.