Ensure AmIUsing renders even without internet access
Previously the CSS blocked rendering, which meant that if it went slowly
(e.g. if we had an invalid proxy connection) then the page was simply
blank indefinitely. Instead, we now move the link into the end of the
body (valid in HTML5, seems supported in FF & Chromium at least) so it
doesn't block rendering, and we use a slightly shorter font-display to
avoid waiting for fonts if it's slow.

This will affect free users immediately, but it won't affect Pro users
until their rules next reset, since the default rules aren't actively
replaced automatically.