Include only IPv4 addresses in Android setup
Only IPv4 capture & rewriting is supported in the Android VPN app, so
using an IPv6 address on a device where that's even supported results
in completely invalid packets and breakage. In addition, we're tight
on space in this QR code anyway, so dropping unusable addresses is a
real help for some cases.