Drop content-length from code exports to fix edge cases
In some cases, the content-length we receive won't match the exported
content length. This happens with some semantically equivalent bodies
for example when sending %40 instead of @ in the body. That becomes @
in the HAR, which is then sent literally (because encoding is not
necessary) in the code, and so the length is different. Instead, we
now delegate this to the given tooling to handle appropriately.

Tested with Curl, Python, Ruby & Node. Works fine in all, although
Node uses chunked encoding by default instead.