Avoid broken UI when editing headers with duplicate names
Previously, if you entered a header name that matched an existing row,
your row was reordered to join the previous row, without moving your
focus, and everything basically broke. This happened because the stored
form of edited headers ignored order, so that data was lost in the 2-way
binding.

To fix this, EditableHeaders and EditablePairs can now be given an input
transform (which should be visible to the user as they type) and a
result conversion callback (which can be used to convert the result, and
detect whether a new input is semantically equivalent to existing
internal state).

By using this, the duplicated logic handling this elsewhere can all be
moved into EditableHeaders/Pairs, and we guarantee that future uses will
do the right thing automatically.

One remaining awkward case: HeaderMatcher uses a format that's even more
lossy, collapsing headers into a single value. Shouldn't cause immediate
problems, needs a breaking change in Mockttp to let us match against raw
headers to be able to deal with this properly.