Drop unnecessary node-dev dev dependency
It connects to a minor (not relevant) marked vulnerability, which is
annoying, but it's also not really useful. It was here to reload the
server if the server content changed (for server+UI development) but
whtn you're developing both together you should use start:web and run a
server separately (the server has its own autoreload dev script) so this
isn't really useful.