Fix rule-reloading when a buffer-backed static rule is present
Previously, these rules were serialized correctly (because the data is a
Buffer, which sets toJSON) but they weren't reloaded when you simple
called serialize->deserialize, because the raw pre-JSON format was a
plain UInt8Array (in some cases? all cases?) and this isn't supported by
the deserialization, which always expects JSON-format data.

This fixes that, by correcting the serialization to proactively
serialize in the toJSON SerializedBuffer format straight away.