Downgrade @reach/router to fix view list bug
A previous minor upgrade changed behaviour, so that the view page was
entirely remounted when moving from /view to /view/:id. This meant that
the first click on a row didn't properly focus the row & reset the list
scroll entirely. This behaviour changed in v1.3.0.

This upgrade wasn't necessary - it just resolved some deprecated dep
warnings that don't affect us, so it's worth rolling back now to fix
that.

In future we should probably move to react-router, but that's
non-trivial since we do some funky things with history & navigation, and
because it is very invested in hooks (while we're decidedly not, and
mobx makes them difficult in some places).

The original commits that updated this, for reference, were:
* https://github.com/httptoolkit/httptoolkit-ui/commit/5727d6d43d67f281b92a9fcdefbe1adca4be226a
* https://github.com/httptoolkit/httptoolkit-ui/commit/d70a2fb8302415d3bfb3ec7d545b3a730e10ca99