Improve click performance in long WebSocket streams
Most important part is the collapsed row onClick. Previously this was a
fresh function every time, so a fresh render. Now, we have a fixed
function and the row passes up its own index. We use memo to skip the
row itself given that, and rendering for ~1000 rows drops from 200ms to
15ms.

There'll still be performance issues with much longer streams (in many
ways) but we'll ignore those for now.