Avoid loading spinner for short-ish JSON & Hex content
For small inputs, render time is ~1ms, so it's not worth passing out to
the background worker. This isn't true for much larger inputs
(background worker is slow in that case too, but at least it's largely
async and we can provide some feedback in the meantime).

This is especially important for WebSockets, because they frequently
have many very short JSON messages, and the flash of spinner every click
is not good UX.