Fix header layout glitch in body cards during decoding
Previously we accidentally layered two <header> elements, not just one.
The flexbox styling for card headers is set with "> header" which means
it wasn't applied to the internal header, and so the header was split
over multiple lines and the layout jumped a little.

That's now fixed, so the layout is consistent. There's still a small
flicker, but only because the size only jumps after the content has been
decoded (even if the card is collapsed). Note that formatting _doesn't_
happen if the card is collapsed, we only decode compressed data to get
the real content size.