Disable Electron on Mac entirely, reverting part of 43a449a
Unfortunately our quick workaround depends on prompt(), which isn't
actually available in Electron regardless, so it doesn't work their
either.

This is tricky, we aren't going to find a pleasant quick fix, and it's
currently breaking things in weird ways for every Mac user that tries
it. Best to explicitly disable it as unsupported instead.