Refactor rule-part availability separate from description definitions
Previously, we allowed the description methods to return undefined, and
anything undefined was hidden in the list. This is not a great way to
encode that information. Similarly, the selection pickers themselves
included logic to check for version support for individual rule parts.

Now, there's a single method for both matchers and handlers that defines
this explicitly in one place, so we can manually enable/disable them
with our own rules directly.