Block all use of the app by banned users
This is sad, but it's the only way to stop some substantial abuse of
users who keep signing up for subscriptions, using the app, and then
disputing the charges with their banks (resulting in fines for HTTP
Toolkit, and risking a permanent ban from payment providers). These
users are either stealing credit cards, from innocent users elsewhere
who report the charges, or they're using their own cards and lying to
the bank to say that they don't recognize the charge to avoid paying
for software they've purchased and used.

Either way, it seems legitimate to block them entirely. Please don't
file chargebacks, they're an existential risk for the tool for
everybody, and I'm quite happy to give refunds to anybody who isn't
satisfied with Pro.