Ensure incompatible UI for new matchers disappears on rule type changes
Previously, if you picked a matcher type, the config would show, and
would stay there even if the type was no longer available in the
dropdown, e.g. because the available matchers had changed because the
rule type had changed, and the matcher you'd selected should no longer
appear. Now that incomplete config resets whenever the type changes,
which should clear that up.

In future we could make this more lax, so that it only resets if the new
type is incompatible, but that's quite a bit more complicated and maybe
not very necessary.