Better handle a UI failing to set rules the server doesn't support
This can only happen (AFAIK) if a Pro user configures rules in the UI,
which persist, and then next run an older server is used, and so the
rules fail to load.

This could also happen in future if we allow configuring a rule in the
UI for versions of the server where it is not supported. But in theory
of course, we'd never do that.

In either case, previously the server would reject the rule
configuration request, and startup would hard fail. Now, the server
will reject it, a warning will be shown & the rules will reset, and
then the app will continue with the reset default rules.