Fix body matching on incomplete requests
This logic was subtly wrong, highlighted by a separate Mockttp bug which
stopped some requests being considered as fully completed.

In fact, you can match the body of an incomplete request - it's totally
possible for a request to be ongoing (client hasn't finished sending its
body) while the server has already sent a response body. Clients should
usually stop at that point, but AFAIK they're not required to, and it's
certainly possible to write a client that doesn't.