Fix search filter bug with duplicate headers
Previously, the header values were an array, which generated weird
suggestions ("a,b") for headers with values a & b, and crashed if you
actually entered a matching value, because the suggestion value must be
a string.

This fixes that, by explicitly handling the array case when generating
header suggestions, so each value is suggested and handled separately.