Prefetch the checkout URL during login time, to speed up checkouts
This doesn't do a proper prefetch, because later on we will open up the
checkout URL in your system browser, that doesn't share its cache, so we
can't do proper prefetching. That said, querying the API like this in
advance lets it talk Paddle early to build the checkout URL, which saves
a valuable little chunk of time, and potentially make us mildly more
resistent to Paddle issues (it implies a free retry) so this is a nice
little optimization.