Auto-highlight (and now correctly remove) problems in read-only editors
This pops open any known problems if detected in a read-only editor,
when the editor is created (with some careful tricks to stop Monaco from
making the editor take focus when it does so). This change pointed out
a separate existing bug, in that open problem details were not cleared
when the body editor content changed. That's now reset too.