Remove very expensive & unnecessary mobx computed wrappers
Neither of these is necessary. For starters, both are observable refs,
and so will change when updated.

For encodedBody, it's actually an observable promise within the ref,
so structurally checking it definitely achieves nothing. Downstream
usage waits for the returned promise explicitly anyway too.

For decodedBody, the struct here was _very_ expensive (10+ seconds of
comparison on very large bodies like 25MB+). It's also not necessary
because of the explicitly updated observable ref - so all updates are
observable. This does mean we may fire extra updates here in the case
where Monaco fires a change event but the content hasn't actually
changed. That seems very niche, I can't reproduce it, and it's probably
much cheaper than this comparison anyway.