Make the relaxed TLS config description a bit more general
Still doing some testing, but it appears that with Node v18+ it's no
longer possible to support very old TLS versions that we supported
before. Unfortunately v16 is EOL imminently so there's not much option.
For now, we'll leave this description generic and see if anybody does
indeed need this old-TLS fallback support.