Add support for trusting ECDSA CA certs
This is already supported interally (by Node, which just passes this to
OpenSSL) but isn't supported by Forge which is used here for initial
parsing & validation. To work around that, we watch for that specific
error and fallback to our own rough & ready parsing instead. This isn't
especially critical logic, so no big deal if we lose a little strictness
here - the runtime network behaviour will end up the same regardless.