Rename reportError to logError to avoid new global type
There's a relatively recent addition to the DOM global types, due to a
real method available for web workers, which means it's very very easy
to accidentally call reportError without importing it, which fails
everywhere except web workers, where it merely does completely the wrong
thing.

logError is exactly the same, but without that problem.