Fix update worker issues due to lack of old path redirect
Old UIs were still querying the old path (/update-worker.js) to refresh
their service worker, which no longer exists. The new path works just
fine, and is referenced correctly from the HTML, but wasn't accessed
until _after_ the UI had updated, leaving us in a bit of a bind.

This unties that knot with a proper redirect. We should probably leave
this in place for quite a while, since it's required for updates until
all UIs have reached this commit or later.