Avoid race condition in SW version request
Previously the SW could receive the wrong auth token (by reading before
it was written by the main thread) and then loop attempting to
use it unsuccessfully, indefinitely.

Now, the SW reloads the auth token for each test, meaning that
although initial requests may fail, it will recover as soon as the UI
shares the required token.

This also remvoes the 'force update' SW logic, added in the distance
past to handle a mismatch between desktop versions. The desktop version
that requires this is before the minimum version in our runtime config,
so those users will already be prompted to upgrade, and have been for a
very long time.