Add a fallback syntax error warning
Our normal outdated update warning (in checkForOutdatedComponents) won't
actually run if the desktop is so outdated that it can't ever parse the
code itself (this is coming soon, due to ?. optional chaining usage in
many places, which won't work in very old (<v1.1.0) desktop apps.

These apps have already been seeing hard warnings every startup for at
least 6 months, and the installs are at least 4 years out of date. In
total, this makes for approx <0.03% of users in the last 90 days.