Ignore ResizeObserver warnings in webpack dev overlay
These suggest a loop _might_ be happening, but not necessarily that it
actually is, and they trigger during editor layout sometimes (e.g. on
Send page) but the overlay makes them very annoying. This ignores them
entirely. They don't seem to appear in the console, but other uncaught
errors do, so I think that's probably fine.