Fix auto-sync after deserialization of Send request models
Previously, the sync was set up in the constructor (run automatically
with no args) and then the body was replaced by the deserialized
version, so the sync didn't actually work after the first load.

Now the factory is used to create a correct & complete model first, and
automatic deserialization for the body is skipped, so everything's set
up properly instead.