Fix bug where Send header mutations weren't persisted
Direct changes which replaced the headers entirely worked, but mutations
didn't, because we used the raw value without serialization, which meant
the direct equality of the headers object meant that persistence didn't
spot the change. Migrating to proper serialization fixes that, and
moving EditableBody creation into the constructor ensures that that has
a way to look up the real deserialized headers, rather than accessing
the raw unserialized data.