Fix isProbablyUtf8 to handle multibyte chars over the 1024 byte line
Previously, multibyte chars crossing this line would make a valid UTF-8
string be considered as binary. That's particularly bad because if
you're modifying a string containing multibyte chars it might change
from utf-8 to binary when a multibyte char is pushed across that line,
resulting in a very weird edit experience.