Update Mockttp for HTTP/2 & trailer improvements
Most of these are server side, but notably this makes it possible for us
to (eventually) listen for trailers and include those in the inspectable
UI data too. For now, we do enough to make the types work, but we don't
actually subscribe to these events or show trailer data yet.