Don't show perf & export cards for invalid (client error) requests
Without this, it's possible to send an invalid request, that will then
generate a code export with an invalid (and in some cases potentially
dangerous, since httpsnippet assumes header keys are valid when
escaping) code snippet.

There's no good reason to export code that will send a totally
unparseable request, and any perf information is probably invalid anyway
since the request hasn't completed the normal request processing etc
(this is more debateable, but personally I'd be surprised if many users
are interested in the performance of invalid requests).