Fix webpack prod build on Windows
For some reason, building the fonts into a subdirectory breaks the
integrity transform in the workbox plugin, which fails the build, but
only on Windows. Presumably something to do with path formats?
Regardless, we're changing the font setup soon, and this solves it for
the short term without consequence, so let's fix it in the easiest
possible way.