Switch to using ... marker for long headers as expand/collapse button
Previously we used the +/- for the whole row (which shows the docs) but
this gives a bit more control, and seems to be what people expect (a few
reports of users confused by previous behaviour). This PR also increases
the length where this kicks in and makes the gradient more obvious to
try to make it clearer what's going on.