Fix overflow issue on Send page body editor
Previously, when the Monaco suggestion window popped up, it used an
absolute layout that resulted in a broken scrollbar on the editor if it
went off the edge of the container.

This options changes the layout so it now works correctly, popping out
of the area entirely as a floating tooltip as expected.