Big performance boost for event data processing
This is most notable under bursts of traffic or when importing a HAR -
in the pathological case, this provides performance boosts of about 60x.
This works by avoiding complete regeneration of the eventStore.exchanges
etc filtered lists (now manually adding/removing new items, instead of
recalculating the list) and using an id to event index (also with
efficient recalculation) to find existing exchanges quickly without
needing to scan the list.