Additional 30% memory reduction by simplifying API data observability
With this change, we no longer use a lazy observable promise for this,
and we no longer instantiate the promise at all before it's required,
and we fix an issue where the list itself triggered API lookup for all
visible rows.

This drops the per-exchange memory overhead by about 3.7KB in some test
examples.