Update to Mockttp v4 for SOCKS, non-HTTP, delays & URL transforms
Lots of exciting things here! None of it directly exposed yet - this PR
just updates and does the minimum required to match breaking changes &
renames etc.

Note that this bumps the minimum desktop & server versions, to versions
which don't need the polyfill (though we include it anyway to give a
little slack there). Should affect <0.3% of users.