Ensure Monaco layout() calls defer until cleared content is be unloaded
Previously, unloading an editor full of content could be expensive. My
beset guess here is that this was caused by Monaco being unmounted
(moving into a world where it was effectively tiny) but not yet updating
the content itself, and then relayout this being extremely expensive
(imagine calculating line wrapping on a 1px x 1px square). When
clearing, we now defer the relayout slightly, which seems to reliably
avoid this case and give us an instant relayout instead.