Size body card as a percentage of container space, with no fixed max
Previously it was fixed to a max of 560px in all cases. This is still
uesd elsewhere (e.g. rule editors) but now in body cards, in modern
desktop versions, we use container queries to size the body to 70% of
the available space (as a max - smaller is fine). This makes it easier
to see more of the important content, without risking very awkward UI
cases by being larger than the space available (especially in the narrow
layout).